gateway: clone ResultProxy at gateway boundaries
Previously, a frontend returning two identical references caused a
double-release error, as the gateway expects each ref to be unique.
However, there are scenarios where a frontend might reasonably be
expected to provide identical refs, e.g. for 2 platforms that have
binary compatibility, or for the upcoming attestations use-case.

To solve this issue, we refactor the ResultProxy implementation to
support cloning. Since the underlying Result supports reference
counting, we can avoid needing to add another layer of refcounts, and
simply clone the underlying refs.

Additionally, we then use the new Cloning functionality in the gateway
and forwarder (e.g. dockerfile) frontends, and ensure that the same
ResultProxy is never returned twice in a result. This means that
identical refs over the wire are translated into the correct
shared-memory structures.

Signed-off-by: Justin Chadwell <me@jedevc.com>