generated-files: add protobuf attestations support
Introduce attestations message structures to protobuf. Each result can
contain multiple attestations, keyed by the platform they apply to.
Each attestation has a predicate type, a ref+path to the predicate
content, and a selection of subjects, which can either be freeform
name + digest fields, or the resulting image manifest for the target.

Signed-off-by: Justin Chadwell <me@jedevc.com>