attestations: move extract logic to utils package
This allows the same logic to be reused by other exporters later.

Additionally, the bundling logic is slightly reworked, to add an
explicit Unbundle step before Extracting. This ensures that the Extract
method can perform a one-to-one translation from result.Attestations to
intoto.Statements, which allows for deriving intermediate properties
such as the Path.

Signed-off-by: Justin Chadwell <me@jedevc.com>